+++ src/main/groovy/org/codenarc/rule/formatting/ClassEndsWithBlankLineRule.groovy
@@ -24,7 +24,7 @@ import org.codenarc.rule.Violation
 import org.codenarc.util.AstUtil
 
 /**
- * Check the location of the closing brace of a class. By default, it enforces that there must be a blank line before
+ * Check whether the class ends with a blank line. By default, it enforces that there must be a blank line before
  * the closing class brace, except if the class is empty and is written in a single line. A blank line is defined as
  * any line that does not contain any visible characters.
  * This rule can be configured with the following properties: